search criteriiaand status functionality

+++ property-management-backend/src/main/java/com/waa/PropertyManagment/service/impl/Propertyserviceimpl.java
@@ -1,13 +1,14 @@
 package com.waa.PropertyManagment.service.impl;
 
-import com.waa.PropertyManagment.domain.Property;
-import com.waa.PropertyManagment.domain.Status;
-import com.waa.PropertyManagment.dto.PropertyDto;
+import com.waa.PropertyManagment.entity.Property;
+import com.waa.PropertyManagment.entity.Status;
+import com.waa.PropertyManagment.entity.dto.PropertyDto;
+import com.waa.PropertyManagment.entity.dto.PropertySearchDao;
+import com.waa.PropertyManagment.entity.dto.SearchCriteria;
 import com.waa.PropertyManagment.repository.PropertyRepository;
 import com.waa.PropertyManagment.service.PropertyService;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
@@ -15,6 +16,9 @@ import java.util.List;
 
 @Service
 public class Propertyserviceimpl implements PropertyService {
+    @Autowired
+    PropertySearchDao propertySearchDao;
+
     @Autowired
     private PropertyRepository propertyRepository;
     @Autowired
@@ -55,9 +59,54 @@ public class Propertyserviceimpl implements PropertyService {
     }
 
     @Override
-    public void deleteProperty(Long id) {
-        propertyRepository.deleteById(id);
+    public List<Property> propertiesByAddress(String city) {
+        List<Property> properties = new ArrayList<>();
+        for (Property p : propertyRepository.propertiesbyaddress(city)) {
+            properties.add(p);
+        }
+        return properties;
     }
+
+    @Override
+    public void cancelContingency(Long id) {
+        updateToAvailable(id);
+    }
+
+    @Override
+    public Property updateToPending(Long id) {
+        Property property = propertyRepository.findById(id).get();
+
+        property.setStatus(Status.PENDING);
+        propertyRepository.save(property);
+        return property;
+    }
+
+    @Override
+    public Property updateToAvailable(Long id) {
+        Property property = propertyRepository.findById(id).get();
+        property.setStatus(Status.AVAILABLE);
+        propertyRepository.save(property);
+        return property;
+    }
+
+    @Override
+    public Property updateToCONTINGENT(Long id) {
+        Property property = propertyRepository.findById(id).get();
+        property.setStatus(Status.CONTINGENT);
+        propertyRepository.save(property);
+        return property;
+    }
+
+    @Override
+    public void deleteProperty(Long id) {
+        Property property = propertyRepository.findById(id).get();
+        Status status = Status.PENDING;
+        if (!property.getStatus().equals(status)) {
+            propertyRepository.deleteById(id);
+        }
+        System.out.println("property is pending can not be deleted");
+    }
+
 
     @Override
     public Property updateProperty(Property property, Long userId) {
@@ -81,4 +130,17 @@ public class Propertyserviceimpl implements PropertyService {
         }
         return toUpate;
     }
+
+    @Override
+    public List<Property> findByAllCriteria(String propertyType, String city, String status, Double price, Integer numberOfRooms) {
+        var dtoSearchRequest=new SearchCriteria();
+        dtoSearchRequest.setPropertyType(propertyType);
+        dtoSearchRequest.setCity(city);
+        dtoSearchRequest.setStatus(status);
+        dtoSearchRequest.setPrice(price);
+        dtoSearchRequest.setNumberOfRooms(numberOfRooms);
+        return  propertySearchDao.findAllByCriteria(dtoSearchRequest);
+    }
+
+
 }
